Use CGFloat and macros for various math functions for floats
git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@15142 0fceea05-a30d-0410-8a8b-80ef821fb0a1

+++ BDSKBackgroundView.m
@@ -62,8 +62,8 @@
 }
 
 - (void)drawRect:(NSRect)rect{
-	float blur = 4.0;
-	float offset = 2.0;
+	CGFloat blur = 4.0;
+	CGFloat offset = 2.0;
 	NSSize size = [self bounds].size;
 	NSRect viewRect = NSMakeRect(blur, blur + offset, size.width - 2 * blur, size.height - blur - offset);
 	NSShadow *aShadow = [[[NSShadow alloc] init] autorelease];
